chore: fix utils packages paths
